Stop relying on GitHub cache to validate APIs (#5618)
* Stop relying on GitHub cache to validate APIs

* Remove workflow_dispatch

(cherry picked from commit 0f1657eb932eab5655112eb6efe40e2e745929a5)

# Conflicts:
#	.github/workflows/validate-apis.yml
